Fixing tests. #116
Former-commit-id: 21a3b7a3883c72e6d3dace28481b8e46b9ed7696

+++ src/main/java/ooga/model/actions/moveractions/Gravity.java
@@ -9,16 +9,16 @@ import org.apache.logging.log4j.Logger;
 public class Gravity implements MoverAction{
 
   private static final Logger LOG = LogManager.getLogger(Gravity.class);
-  private final int GRAVITY_VELOCITY;
+  private final double GRAVITY_VELOCITY;
 
   /**
    * Bounce constructor initializes default Bounce velocity value
    */
   public Gravity(){
 
-    int tempVelocity;
+    double tempVelocity;
     try{
-      tempVelocity = Integer.parseInt(
+      tempVelocity = Double.parseDouble(
           ResourceBundle.getBundle("properties/movement").getString("gravity_velocity"));
     }
     catch(NumberFormatException exception){
